Fix live preview getting out of sync
This bug was deeply rooted in how the VCS looks up deltas at an absolute
time. The issue was that we need to first find the chunk whose absolute start
time is < to a time t and then we look for the delta in in that chunk.

However when t was right in between the beginning of the chunk and the padding
of its first delta we would not find a match and end up with a nil
skimhead-entry, which ultimately caused live-preview to start tracking the
playhead instead.

The fix here ensures that when we look for the nearest chunk to the left we also
check that t falls after the first delta, otherwise we'll get the next chunk to
the left where we'll find our delta entry in the last position.