Delta store: remove user-uid from partition key
We had [user-uid vims-uid] as a partition key, but this was unnecessary because:

- vims-uid is enough to partition the deltas
- we never need to grab a user's deltas across multiple vims in one query

On the other hand this was preventing us from adressing a major use case, which
is to retrieve the deltas without knowing the vims owner.

For example when a visitor loads a vims, we were not able to request the deltas
without first going to datomic to find out who the vims owner was.

This was "not a problem" in v1 because the backend was aggregating the data from
both data stores before returning the reponse, but this was very limiting
because:

- we weren't able to return any vims data until we fetched the deltas
- we had to keep all the deltas in memory until the response had transfered

This change will make it possible to display the vims information as the client
loads the deltas, and it will allow the backend to stream the deltas
asynchronously straight from cassandra.