Use reagent/current-component as remote status-key
This seems to work well for the purpose of a component waiting on a single
remote action. The component, and thus the key, doesn't change until the
response comes back, and when the component updates after the callback the
status key changes (the component is now a different object) so we don't get a
stale status.

This pattern is likely to cause issues in a component that subscribes to two or
more subs with remote fxs and their statuses. The first status to succeed will
likely cause the component to re-render which will "reset" the other ones.